players scripts v0.8
My twitch script to see how many players are in your match at the beginning of the match. 
All detail for installing it are in the README.txt. If you have any issue @ me :) 
I'll improve it later to see the name of the peoples in the game and update more data but that's already good for now