1.9 HUGE update (kinda)
Add stats of all the class you encounter, to be able to say that tox and temp are overplayed
Do a database with the name of all the people you encounter. So if you suspect someone to have changed his name you can see all the name he used against you.
Add some cooldown on commands (to prevent spam)